refactor(pvtpkgs): convert to subflake
Main reason is so that I can specify a flake input as a package src, so
I don't have to manually update packages. Will be useful for vim
plugins.

Also, seems like the lock file entry for the subflake is going to be
changed every rebuild... `https://github.com/NixOS/nix/pull/6530` will
likely fix this.

Signed-off-by: Dixon Sean Low Yan Feng <dixonseanlow@protonmail.com>